fix so works better
- This is better but needs more testing/work.
However, it is time to commit.
- Make so knows if production install and
acts accordingly.
- The password check function now returns if
the password should be changed rather than
if it generated a new one so it works in
all cases.
- Put all three environment variables that
can change in .env not just ones changed.
This seemed better since you cannot change
values in the docker config after the initial startup
so now all are in that one file.
- Fixed bug where did not change default
password if it was not changed. Should not
happen once everything works but safer to
make it the first check.
- Added the postgres password to the common
environment variables in docker-compose.yml
since needed by install during web process
to get an password updates.
- Update npm script call to change passwords
so new name and directly calls the function.
This avoids the default running since that
can cause errors if improperly used
and it may not be supported in the future.
- Changed the password change file name to
match script and reflect new usage.
- There are debug statement left in for now.
